Don't need time.sleep. Don't need page locators. Locators are in they pages
