+++ pages/history_page.py
@@ -1,4 +1,6 @@
 from .base_page import BasePage
+from selenium.webdriver.common.by import By
+
 
 class HistoryPage(BasePage):
