+++ test_main_page.py
@@ -12,10 +12,8 @@ link = "https://www.selenium.dev/"
 def test_open_documentation(browser):
     page = MainPage(browser, link)
     page.open_page()
-    time.sleep(2)
     page.should_be_link_to_document_page()
     page.go_to_documentation()
-    time.sleep(2)
     # Проверяем что это точно страница Documentation
     page = DocumentPage(browser, link)
     page.should_be_document_page()
@@ -25,10 +23,8 @@ def test_open_documentation(browser):
 def test_open_readmore_webdriver(browser):
     page = MainPage(browser, link)
     page.open_page()
-    time.sleep(2)
     page.should_be_readmore_btn()
     page.go_to_readmore()
-    time.sleep(2)
     # Проверяем что это точно страница WebDriver
     page = WebdriverPage(browser, link)
     page.should_be_webdriver_link()
@@ -37,10 +33,8 @@ def test_open_readmore_webdriver(browser):
 def test_open_history(browser):
     page = MainPage(browser, link)
     page.open_page()
-    time.sleep(2)
     page.should_be_link_to_about_menu()
     page.go_to_about_menu()
-    time.sleep(2)
     #Переходим на страницу History
     page.should_be_link_to_history()
     page.go_to_history()
@@ -55,20 +49,15 @@ def test_open_webdriver_page(browser):
     page = DocumentPage(browser, link)
     # создает экземпляр страницы Documentation
     page.open_page()
-    time.sleep(2)
     page.go_to_webdriver()
     page = WebdriverPage(browser, link)
-    time.sleep(2)
     page.should_be_webdriver_link()
 
 # Тест проверяет, что пользователь может ввести слово в поиск и перейти по нему
 def test_should_use_search(browser):
     page = MainPage(browser, link)
     page.open_page()
-    time.sleep(2)
     page.should_be_search_btn()
-    time.sleep(2)
     page.use_search(search='grid')
-    time.sleep(10)
